+++ src/pages/home/HomePage.tsx
@@ -1,11 +1,21 @@
+import { getRole, getToken } from '@utils/storage';
+import { useEffect } from 'react';
+import { useNavigate } from 'react-router-dom';
+
 const HomePage = () => {
-  return (
-    <div>
-      <h1>GUEST일 경우</h1>
-      <h1>SENIOR일 경우</h1>
-      <h1>JUNIOR일 경우</h1>
-    </div>
-  );
+  const navigate = useNavigate();
+  const token = getToken();
+  const role = getRole();
+
+  useEffect(() => {
+    if (token && role) {
+      navigate(role === 'SENIOR' ? '/promiseList' : '/juniorPromise');
+    } else {
+      navigate('/join');
+    }
+  }, [token, role]);
+
+  return null;
 };
 
 export default HomePage;
